+++ src/buildConfig.js
@@ -6,6 +6,7 @@ import {
 import { configurationError } from "./utils"
 import cosmiconfig from "cosmiconfig"
 import fs from "fs"
+import globjoin from "globjoin"
 import path from "path"
 import resolveFrom from "resolve-from"
 
@@ -99,9 +100,9 @@ export default function (options) {
  * @param {string} [options.ignorePath] - See above.
  * @return {object}
  */
-export function augmentConfig(config, configDir, options = {}) {
+function augmentConfig(config, configDir, options = {}) {
   const start = (options.addIgnorePatterns)
-    ? addIgnoreFiles(config, configDir, options.ignorePath)
+    ? addIgnores(config, configDir, options.ignorePath)
     : Promise.resolve(config)
   return start.then(configWithIgnorePatterns => {
     const absolutizedConfig = absolutizeExtras(configWithIgnorePatterns, configDir)
@@ -112,31 +113,33 @@ export function augmentConfig(config, configDir, options = {}) {
   })
 }
 
-export function addIgnoreFiles(config, configDir, ignorePath) {
-  return findIgnorePatterns(configDir, ignorePath).then(ignorePatterns => {
-    config.ignoreFiles = [].concat(ignorePatterns, config.ignoreFiles || [])
-    return config
-  })
-}
+function addIgnores(config, configDir, ignorePath) {
+  // Absoluteize config.ignoreFiles
+  if (config.ignoreFiles) {
+    config.ignoreFiles = [].concat(config.ignoreFiles).map(glob => {
+      if (path.isAbsolute(glob)) return glob
+      return globjoin(configDir, glob)
+    })
+  }
 
-function findIgnorePatterns(configDir, ignorePath) {
-  let defaultedIgnorePath = path.resolve(configDir, IGNORE_FILENAME)
+  let defaultedIgnorePath = path.resolve(process.cwd(), IGNORE_FILENAME)
   if (ignorePath) {
     defaultedIgnorePath = (path.isAbsolute(ignorePath))
       ? ignorePath
       : path.resolve(process.cwd(), ignorePath)
   }
+
   return new Promise((resolve, reject) => {
     fs.readFile(defaultedIgnorePath, "utf8", (err, data) => {
       if (err) {
         // If the file's not found, great, we'll just give an empty array
-        if (err.code === FILE_NOT_FOUND_ERROR_CODE) { return resolve([]) }
+        if (err.code === FILE_NOT_FOUND_ERROR_CODE) { return resolve(config) }
         return reject(err)
       }
-      const ignorePatterns = data.split(/\r?\n/g)
-        .filter(val => val.trim() !== "")     // Remove empty lines
-        .filter(val => val.trim()[0] !== "#") // Remove comments
-      resolve(ignorePatterns)
+      // Add an ignorePatterns property to the config, containing the
+      // .gitignore-patterned globs loaded from .stylelintignore
+      config.ignorePatterns = data
+      resolve(config)
     })
   })
 }